Move content to components folder
